Add files via upload
main.py
modified the delete if expired function so it wont require channel id to run. this was to fix issues with over iterating and overlapping loops (might still not be resolved)

imger.py
updated to include cyrillic and russian.
make sure the tesseract installed with all the languages if can for future possibilities. needs to include russian languages on installation